real_robot_pipeline/requirements.txt
# Core dependencies
numpy>=1.20
pillow>=9.0
pyyaml>=6.0

# UI
gradio>=4.0

# Model inference
lerobot[cuda]>=0.1.0
jax>=0.4.0
jax-cuda11-cudnn82>=0.4.0

# Robot control
pyserial>=3.5

# ROS (optional)
rospy
geometry_msgs

# Development (optional)
pytest>=7.0
black>=22.0
